Move SEO component in Layout
